[dnssec-keymgr] add directory option to policy
A directory can now be specified in a policy definition.  If zones
are specified on the keymgr command line and -K isn't, then the zones'
policy-defined directories are used. if zones are not specified, then a
specified directory will be searched for keys, as before.

Some changes were also made to dnssec-coverage to adjust to the new
keydict functionality.